Minor changes to storage.py
Created new build.py file that contains a slew of transactions and methods for configuring a UCS domain from scratch.
These were captured using the ConvertTo-UcsCmdlet PoSH cmdlet and as a result, there will be many improvements to the code provided here.
It may not even stay in this library for long...just committing it here so it can have a home while I work on it.